+++ src/client/components/Document/styled.ts
@@ -5,6 +5,7 @@ export const StyledDocument = styled.main`
   height: calc(100vh - var(--header-height));
   display: inline-block;
   overflow-y: auto;
+  position: relative;
 
   @media (max-width: 768px) {
     width: 100vw;
@@ -15,11 +16,14 @@ export const StyledDocument = styled.main`
 export const StyledMarkdown = styled.div`
   color: var(--c-1);
   padding: 0 4em;
+  padding-right: calc(4em - 20px);
   word-break: break-all;
   margin-bottom: 100px;
+  width: calc(100% - var(--toc-width));
 
   @media (max-width: 768px) {
     padding: 0 1em;
+    width: 100%;
   }
 
   h1,
@@ -77,6 +81,7 @@ export const StyledMarkdown = styled.div`
   }
 
   p,
+  .docit-paragraph,
   ol,
   ul {
     margin: 1rem 0;
@@ -174,6 +179,10 @@ export const StyledMarkdown = styled.div`
   }
 
   blockquote > p {
+    margin: 0;
+  }
+
+  blockquote > .docit-paragraph {
     margin: 0;
   }