+++ src/client/shim.d.ts
@@ -28,6 +28,17 @@ interface Overrides {
   style?: React.CSSProperties;
 }
 
+interface ParsedTocItem {
+  title?: string;
+  url?: string;
+  items?: ParsedTocItem[];
+}
+
+interface Markdown {
+  routePath: string;
+  toc: ParsedTocItem;
+}
+
 declare module "virtual:routes" {
   const routes: Route[];
   export { routes };
@@ -39,7 +50,12 @@ declare module "virtual:appData" {
     twitter?: string;
     github?: string;
   };
-  export default { title, socials };
+  const version: string;
+  const markdowns: {
+    routePath: string;
+    toc: ParsedTocItem;
+  }[];
+  export default { title, socials, version, markdowns };
 }
 
 declare module "virtual:sidebars" {