+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.7.0] - Unreleased
 ### Added
 - Not web console will print out docit version
+- Add `TOC` section
 
 ### Fixed
 - sidebar anchors now have hover and active style